processing/pyproject.toml
[tool.poetry]
name = "payscope-processing"
version = "0.1.0"
description = "PayScope processing service"
authors = ["PayScope <dev@payscope.local>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.4.0"
redis = "^5.0.0"
prometheus-client = "^0.20.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
anyio = "^4.6.0"

# Parsing & extraction
pandas = "^2.2.0"
openpyxl = "^3.1.0"
charset-normalizer = "^3.3.0"
unstructured = "^0.16.0"
pymupdf = "^1.24.0"
pillow = "^10.4.0"
paddleocr = "^2.8.0"
lingua-language-detector = "^2.0.0"

# Layout understanding
torch = "^2.3.0"
transformers = "^4.44.0"

# Storage
boto3 = "^1.35.0"
httpx = "^0.27.0"
python-dateutil = "^2.9.0.post0"
pinecone = "^5.0.0"
neo4j = "^5.23.0"
sentence-transformers = "^3.0.1"
langchain-core = "^0.3.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
crewai = "^0.80.0"
pyautogen = "^0.2.0"
trl = {version = "^0.9.0", optional = true}
psycopg = {version = "^3.1.18", extras = ["binary"]}
prophet = "^1.1.5"
networkx = "^3.3"
cryptography = "^42.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"


